Move ome_* and idrac_* modules one level up.
